Refactor GoogleMap component to utilize MapStateLogger for improved event tracking. Removed inline MapStateLogger implementation and added it as a separate component, enhancing code organization and maintainability.
